Control building from feature flag
Properly skips optional dependencies pending rust-lang/cargo#7914
More extensible for addition of new platforms
