+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "glslang"]
-	path = glslang
+	path = glslang_actual_builder/glslang
 	url = https://github.com/KhronosGroup/glslang.git
