+++ Cargo.toml
@@ -5,11 +5,23 @@ authors = ["Pierre Krieger <pierre.krieger1708@gmail.com>", "The vulkano contrib
 description = "Deprecated. This crate is a temporary measure until native rust shader compilation like https://github.com/gfx-rs/naga lands."
 repository = "https://github.com/cart/glsl-to-spirv"
 license = "MIT/Apache-2.0"
-build = "build/build.rs"
 categories = ["rendering::graphics-api"]
 edition = "2018"
-links = "glslang"
 
-[build-dependencies]
-cmake = "0.1.45"
-cc = { version = "1.0.66", features = ["parallel"] }
+[target.x86_64-pc-windows-msvc.dependencies]
+glslang_actual_builder = { path = "glslang_actual_builder", default-features = false }
+
+[target.x86_64-unknown-linux-gnu.dependencies]
+glslang_actual_builder = { path = "glslang_actual_builder", default-features = false }
+
+[target.x86_64-apple-darwin.dependencies]
+glslang_actual_builder = { path = "glslang_actual_builder", default-features = false }
+
+[target.aarch64-linux-android.dependencies]
+glslang_actual_builder = { path = "glslang_actual_builder", default-features = false }
+
+[target.armv7-linux-androideabi.dependencies]
+glslang_actual_builder = { path = "glslang_actual_builder", default-features = false }
+
+[target.i686-pc-windows-msvc.dependencies]
+glslang_actual_builder = { path = "glslang_actual_builder", features = ["build-from-source"] }
