+++ glslang_actual_builder/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "glslang_actual_builder"
+version = "0.1.0"
+authors = ["Pierre Krieger <pierre.krieger1708@gmail.com>", "The vulkano contributors", "Carter Anderson <mcanders1@gmail.com>", "Nicholas Rishel <rishel.nick@gmail.com>", "Martin Krisnanto Putra <grygrflzr@hotmail.com>"]
+description = "Internal builder"
+repository = "https://github.com/cart/glsl-to-spirv"
+license = "MIT/Apache-2.0"
+build = "build.rs"
+categories = ["rendering::graphics-api"]
+edition = "2018"
+
+[features]
+default = []
+build-from-source = ["cmake", "cc"]
+
+[build-dependencies]
+cmake = { version = "0.1.45", optional = true }
+cc = { version = "1.0.66", features = ["parallel"], optional = true }
